requirements.txt
streamlit
google-auth
google-auth-oauthlib
google-api-python-client
langchain
openai
chromadb
python-dotenv
tqdm